+++ projects/start-os/web/ui/src/app/services/api/embassy-mock-api.service.ts
@@ -543,10 +543,10 @@ export class MockApiService extends ApiService {
     return null
   }
 
-  async queryDns(params: T.QueryDnsParams): Promise<string | null> {
+  async queryDns(params: T.QueryDnsParams): Promise<T.QueryDnsRes> {
     await pauseFor(2000)
 
-    return null
+    return { ipv4: null, ipv6: null }
   }
 
   async checkPort(params: T.CheckPortParams): Promise<T.CheckPortRes> {
@@ -558,6 +558,18 @@ export class MockApiService extends ApiService {
       openExternally: true,
       openInternally: false,
       hairpinning: true,
+    }
+  }
+
+  async checkPortV6(
+    params: T.CheckPortParams,
+  ): Promise<T.CheckPortV6Res | null> {
+    await pauseFor(2000)
+
+    return {
+      ip: '::',
+      openExternally: true,
+      openInternally: false,
     }
   }
 
@@ -1605,7 +1617,7 @@ export class MockApiService extends ApiService {
     this.mockRevision(patch)
 
     return {
-      dns: null,
+      dns: { ipv4: null, ipv6: null },
       port: {
         ip: '0.0.0.0',
         port: 443,
@@ -1613,6 +1625,7 @@ export class MockApiService extends ApiService {
         openInternally: false,
         hairpinning: false,
       },
+      portV6: null,
     }
   }
 
@@ -1749,7 +1762,7 @@ export class MockApiService extends ApiService {
     this.mockRevision(patch)
 
     return {
-      dns: null,
+      dns: { ipv4: null, ipv6: null },
       port: {
         ip: '0.0.0.0',
         port: 443,
@@ -1757,6 +1770,7 @@ export class MockApiService extends ApiService {
         openInternally: false,
         hairpinning: false,
       },
+      portV6: null,
     }
   }